use dpram in 2-bit write, 16-bit read for clock crossing
This overhauls the TMDS decoder and clock crossing from the
bit_clk domain to hdmi_clk domain by using a dpram block
in 2-bit write, 16-bit read to act as a clock crossing and
shift-register.

Timing analysis is 133 MHz, which should be in excess of
the 125MHz required for the bit clock, but this requires
explicit placement of the DPRAM's close to the TMDS input
pins.  This is potentially fragile, but seems to produce
the best results.